+++ GitUserApp/GitUserApp/Resource/Assets.xcassets/AppIcon.appiconset/Contents.json
