+++ GitUserApp/GitUserApp/Resource/Assets.xcassets/Contents.json
