+++ GitUserApp/GitUserApp/Resource/Assets.xcassets/placeholder.imageset/Contents.json
