+++ GitUserApp/GitUserApp/Resource/Assets.xcassets/placeholder.imageset/octoface@2x.png
