+++ GitUserApp/GitUserApp/Resource/Assets.xcassets/placeholder.imageset/octoface@3x.png
