+++ GitUserApp/GitUserApp/Resource/Info.plist
